requirements.txt
numpy ~= 1.14.3
scipy ~= 1.0.1
future ~= 0.16.0
matplotlib ~= 2.2.2
pillow ~= 5.0.0
scikit-image ~= 0.14.0
pyaml
